% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{add.lineage.pairs}
\alias{add.lineage.pairs}
\title{add.lineage.pairs}
\usage{
add.lineage.pairs(model, lineage)
}
\arguments{
\item{model}{= MODEL object}

\item{lineage}{= Lineage object to be added to list of lineage pair choices}
}
\description{
this function adds lineage pairs in MODEL obj attr `choices`: list of lineage 
pair choices to coalesce
}
