% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventLogger.R
\name{as.phylo.EventLogger}
\alias{as.phylo.EventLogger}
\title{as.phylo}
\usage{
\method{as.phylo}{EventLogger}(eventlog, transmissions = FALSE,
  migrations = FALSE)
}
\arguments{
\item{eventlog:}{EventLogger object}

\item{transmissions:}{logical; if TRUE, transmission events included, 
else excluded otherwise.}
}
\value{
ape::phylo object
}
\description{
An S3 method converts events stored in the EventLogger object into an inner 
coalescent tree w/ option to include/exclude transmission events.
}
\examples{
path <- system.file('extdata', 'structSI.yaml', package='twt')
settings <- yaml.load_file(path)
structSI <- Model$new(settings)
run <- sim.outer.tree(structSI)
eventlog <- sim.inner.tree(run)
tr <- as.phylo(eventlog, transmissions=TRUE, migrations=TRUE) 
tr

# make a nice plot
ggtree(tr) + geom_tiplab() + 
geom_nodepoint(aes(fill=(tr$event.type=='coalescent')), 
pch=21, stroke=1, size=3, colour='black')

}
