% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InnerTree.R
\name{as.phylo.InnerTree}
\alias{as.phylo.InnerTree}
\title{as.phylo.InnerTree}
\usage{
\method{as.phylo}{InnerTree}(obj)
}
\arguments{
\item{obj:}{R6 object of class `InnerTree`}
}
\value{
object of class `phylo`
}
\description{
A generic S3 method for converting an InnerTree object to an S3 object of 
class `ape::phylo` that represents a phylogenetic tree.  This inner 
phylogeny is nested within an "outer" transmission tree.
}
