% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OuterTree.R
\name{as.phylo.OuterTree}
\alias{as.phylo.OuterTree}
\title{as.phylo.OuterTree}
\usage{
\method{as.phylo}{OuterTree}(obj)
}
\arguments{
\item{obj:}{R6 object of class `OuterTree`}
}
\value{
S3 object of class `phylo`
}
\description{
A generic S3 method for converting an OuterTree to a phylo object.
`ape::phylo` is an S3 class object that represents a phylogenetic tree. We 
use this object class to represent the transmission tree.  Because this 
object is, by definition, an acyclic graph, it cannot represent cases of 
superinfection.  For this reason, only the earliest transmission event is 
used for each recipient host.  To represent the entire transmission 
history including superinfection, use `as.igraph.OuterTree`.
}
\details{
Note that interpreting the resulting transmission tree as a pathogen
phylogeny ignores all within-host evolution, e.g., lineage sorting.
}
