% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run.R
\name{as.phylo.Run}
\alias{as.phylo.Run}
\title{as.phylo.Run
Generic method for Run objects - convert Run object (outer tree 
simulation) into an object of class `phylo`}
\usage{
\method{as.phylo}{Run}(run)
}
\arguments{
\item{run:}{R6 object of class `Run`}
}
\value{
S3 object of class `phylo`
}
\description{
as.phylo.Run
Generic method for Run objects - convert Run object (outer tree 
simulation) into an object of class `phylo`
}
