% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{attach.tree}
\alias{attach.tree}
\title{attach.tree}
\usage{
attach.tree(tree, model)
}
\arguments{
\item{tree:}{either a Newick tree string or an object of class 'phylo' (ape)
that represents the transmission tree (history).  Internal node labels must 
be present to indicate transmission sources.}

\item{model:}{an R6 model of class Model.  The Model should provide the parameters
of a CompartmentType and sampled Compartments.  Currently only one 
CompartmentType is supported by this method, until we can devise a 
procedure for the user to assign unsampled Compartments at internal
nodes to different CompartmentTypes.}
}
\value{
An object of class Model
}
\description{
\code{attach.tree} converts a Newick tree string or ape::phylo object
into a sequence of transmission events that are stored as an eventlog object,
bypassing outer tree simulation
}
\examples{
e <- attach.tree('(((A:1,B:1)B:1,C:1)C:1,D:1)D:1;')
e  # print contents


}
\seealso{
sim.outer.tree, load.outer.tree
}
