% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawCoalWaitTimes.R
\name{calc.coal.wait.times}
\alias{calc.coal.wait.times}
\title{calc.coal.wait.times}
\usage{
calc.coal.wait.times(model, current.time, dynamic = FALSE)
}
\arguments{
\item{model:}{object of class 'MODEL'}

\item{current.time:}{current simulation time for inner tree}
}
\value{
Named numeric vector of waiting times per compartment
}
\description{
Draw waiting times for all compartments that have two or more extant lineages.
}
