% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{Compartment}
\alias{Compartment}
\title{Compartment}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Compartment
}
\arguments{
\item{name:}{a character string that uniquely identifies the Compartment}

\item{type:}{a reference to a CompartmentType object}

\item{source:}{a reference to another Compartment from which a Lineage was 
transmitted to this Compartment}

\item{branching.time:}{stores the origin time of this Compartment, which 
corresponds to a branching event in the "outer" tree.}

\item{unsampled:}{if TRUE, then any Lineage carried by this Compartment is not
directly observed, i.e., it does not represent a tip in the "inner" tree.}
}
\description{
\code{Compartment} is an R6 class for objects that represent the units of an
outer tree simulation, such as a host individual or deme.
}
\examples{
# load Compartments from a YAML object
path <- system.file('extdata', 'SI.yaml', package='twt')
settings <- yaml.load_file(path)
mod <- MODEL$new(settings)

# display first Compartment object
host1 <- mod$get.compartments()[[1]]
host1

# manually initialize a new Compartment object
hostN <- Compartment$new(name='newHost', unsampled=TRUE)
hostN$set.type(host1$get.type())

}
\keyword{datasets}
