% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{.assign.source.compartments}
\alias{.assign.source.compartments}
\title{.assign.source.compartments}
\usage{
.assign.source.compartments(eventlog, comps, source.popns, t_events)
}
\description{
Assign source Compartments for all sampled and promoted unsampled infected
Compartments.
}
\keyword{internal}
