% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{.assign.transmission.times}
\alias{.assign.transmission.times}
\title{.assign.transmission.times}
\usage{
.assign.transmission.times(infected, events, initial.samplings, type)
}
\arguments{
\item{infected:}{named list of Compartment objects to be assigned transmission times}

\item{events:}{list of possible transmission events for the infected}

\item{initial.samplings:}{list of respective initial sampling times for each 
Compartment}

\item{type:}{CompartmentType object}
}
\value{
Returning vector of all transmission times, including times used and unused 
in this current assignment of transmission times.
}
\description{
Assignment of transmission times specific to each type.
}
\keyword{internal}
