% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMigrations.R
\name{.calc.migration.events}
\alias{.calc.migration.events}
\title{.calc.migration.events}
\usage{
.calc.migration.events(init.conds, popn.migration.rates, max.sampling.time,
  possible.source.types, transmission.events)
}
\arguments{
\item{init.conds:}{initial conditions specified in model settings, i.e., 
originTime and size (per CompartmentType)}

\item{popn.migration.rates:}{rates of migration between different 
CompartmentTypes}

\item{max.sampling.time:}{time of the most recent sample as a endpoint for 
this function (type numeric)}

\item{possible.source.types:}{list of possible Sources that each recipient 
type can receive a migrating lineage from}

\item{transmission.events:}{}
}
\value{
data frame of migration events, each made up of: time, 
        recipient Type, and source Type
}
\description{
Generates migration events only, based on population dynamics of the MODEL.
}
\keyword{internal}
