% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{.calc.popn.rates}
\alias{.calc.popn.rates}
\title{.calc.popn.rates}
\usage{
.calc.popn.rates(types, indiv.types)
}
\arguments{
\item{types:}{list of CompartmentType objects}

\item{indiv.types:}{list of CompartmentType name of type `character` for each 
individual Compartment object}
}
\value{
matrix of population transmission rates for each type to type comparison
}
\description{
Stores population rates for each CompartmentType specified by the user.
}
\keyword{internal}
