% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{.calc.transmission.events}
\alias{.calc.transmission.events}
\title{.calc.transmission.events}
\usage{
.calc.transmission.events(popn.totals, popn.rates, init.samplings,
  possible.sources, max.attempts = 10)
}
\arguments{
\item{popn.totals:}{(named matrix) numbers of susceptible and infected 
Lineages at time t=0, specific to each LineageType (rows) and 
CompartmentType (columns)}

\item{popn.rates:}{rates of transmission between different CompartmentTypes}

\item{init.samplings:}{list of first sampling times for each Compartment}

\item{possible.sources:}{list of possible Sources that each recipient Type can 
receive a transmission from}

\item{max.attempts:}{number of tries to sample transmission events}
}
\value{
data frame of transmission events, each made up of: time, 
        recipient Type, source Type, and LineageType (not yet implemented)
}
\description{
Generates transmission events only, based on population dynamics of the
MODEL
}
\keyword{internal}
