% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventLogger.R
\name{.eventlogger.to.phylo}
\alias{.eventlogger.to.phylo}
\title{.eventlogger.to.phylo}
\usage{
.eventlogger.to.phylo(eventlog, transmissions = FALSE,
  migrations = FALSE, node.labels = FALSE)
}
\arguments{
\item{eventlog:}{EventLogger object}

\item{transmissions:}{logical; if TRUE, transmission events included, 
else excluded otherwise.}
}
\value{
ape::phylo object
}
\description{
function converts events stored in the EventLogger object into an inner 
coalescent tree w/ option to include/exclude transmission events.
}
\keyword{internal}
