% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{.get.initial.samplings}
\alias{.get.initial.samplings}
\title{.get.initial.samplings}
\usage{
.get.initial.samplings(compartments)
}
\arguments{
\item{compartments:}{list of sampled infected Compartment objects}
}
\value{
named vector of first Lineage sampling times for each Compartment
}
\description{
Support function for sim.outer.tree.  Retrieves first sampling time of a Lineage 
for each sampled infected Compartment.
}
\keyword{internal}
