% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{.get.rate.matrices}
\alias{.get.rate.matrices}
\title{.get.rate.matrices}
\usage{
.get.rate.matrices(types)
}
\arguments{
\item{types:}{list of CompartmentType objects as returned by Run$get.types()}
}
\value{
list of named matrices of transmission and migration rates for each 
         CompartmentType to CompartmentType
}
\description{
Stores population rates for each CompartmentType specified by the user.
Handle unspecified rate parameters.
}
\keyword{internal}
