% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run.R
\name{.plot.dynamics}
\alias{.plot.dynamics}
\title{.plot.dynamics
Make a stairstep plot of the population dynamics.}
\usage{
.plot.dynamics(run, cex = 1, ...)
}
\arguments{
\item{run:}{R6 object of class Run}

\item{...:}{additional arguments passed to generic plot()}
}
\description{
.plot.dynamics
Make a stairstep plot of the population dynamics.
}
\keyword{internal}
