% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventLogger.R
\name{.plot.outer.tree}
\alias{.plot.outer.tree}
\title{.plot.outer.tree}
\usage{
.plot.outer.tree(e)
}
\arguments{
\item{e:}{EventLogger object}
}
\value{
ape::phylo object
}
\description{
A function that converts events stored in an EventLogger object into
an outer transmission tree.
}
\keyword{internal}
