% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMigrations.R
\name{.record.migration.rates}
\alias{.record.migration.rates}
\title{.record.migration.rates}
\usage{
.record.migration.rates(types, indiv.types)
}
\arguments{
\item{types:}{list of CompartmentType objects}

\item{indiv.types:}{list of CompartmentType name of type `character` for each 
individual Compartment object}
}
\value{
popn.migration.rates = matrix of population migration rates for each 
Type to Type comparison (for all possible migration events)
}
\description{
Helper function stores migration rates for each CompartmentType 
specified by the user.
}
\keyword{internal}
