% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMigrations.R
\name{.record.possible.source.types}
\alias{.record.possible.source.types}
\title{.record.possible.source.types}
\usage{
.record.possible.source.types(types, lineages)
}
\arguments{
\item{types:}{list of CompartmentType objects}

\item{lineages:}{list of Lineage objects}
}
\value{
named list of possible Types of `source` with 
recipient types of type character
}
\description{
Stores list of possible Types of `source` with recipient types for a 
migration event as type character.
}
\keyword{internal}
