% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run.R
\name{.reorder.events}
\alias{.reorder.events}
\title{.reorder.events}
\usage{
.reorder.events(events, node, result = c())
}
\description{
A helper function that recursively generates a list of node labels
by post-order traversal (outputting parents after children).
}
\keyword{internal}
