% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventLogger.R
\name{.reorder.inner.events}
\alias{.reorder.inner.events}
\title{.reorder.inner.events
Recursive helper function to generate ordering of events by preorder
traversal of inner tree.  Similar to Run::.reorder.events().}
\usage{
.reorder.inner.events(events, node, order, result = c())
}
\arguments{
\item{events:}{data frame of core (coalescent, bottleneck) events}

\item{node:}{character, unique identifier of a node}

\item{result:}{vector to pass between recursive calls}
}
\description{
.reorder.inner.events
Recursive helper function to generate ordering of events by preorder
traversal of inner tree.  Similar to Run::.reorder.events().
}
\keyword{internal}
