% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{.resolve.migration}
\alias{.resolve.migration}
\title{.resolve.migration}
\usage{
.resolve.migration(run, e)
}
\arguments{
\item{run:}{R6 object of class Run}

\item{e:}{a row from EventLogger$get.all.events() dataframe}
}
\description{
Helper function records the migration of extant Lineages between
Compartments.
}
\keyword{internal}
