% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{.resolve.transmission}
\alias{.resolve.transmission}
\title{.resolve.transmission}
\usage{
.resolve.transmission(run, e)
}
\arguments{
\item{run:}{R6 object of class Run}

\item{e:}{row from EventLogger dataframe}
}
\description{
Helper function records the transmission of Lineages from source
to recipient Compartments and updates the state of the Run object.
}
\keyword{internal}
