% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{.rexp.coal}
\alias{.rexp.coal}
\title{rexp.coal}
\usage{
.rexp.coal(run, comp, time)
}
\arguments{
\item{run:}{R6 object of class Run}

\item{comp:}{R6 object of class Compartment}

\item{time:}{double, current simulation (reverse) time}
}
\value{
Random variate from waiting time distribution.
}
\description{
Draw exponential waiting time to next coalescent event of Lineages 
within a given Compartment.  Note that if the waiting time exceeds 
this Compartment's transmission event, it will be discarded by 
sim.inner.tree().
}
\keyword{internal}
