% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{.sample.outer.events}
\alias{.sample.outer.events}
\title{.sample.outer.events}
\usage{
.sample.outer.events(
  types,
  init.conds,
  popn.rates,
  init.samplings,
  max.attempts = 10
)
}
\arguments{
\item{types:}{list of CompartmentTypes as returned by Run$get.types()}

\item{init.conds:}{initial conditions, comprising (1) time scale of simulation
(2) initial population size per CompartmentType and (3) Type
of index case}

\item{popn.rates:}{names matrices for transmission, transition and migration
rates.}

\item{init.samplings:}{list of first Lineage sampling times per Compartment,
grouped by CompartmentType}

\item{max.attempts:}{number of tries to sample events}
}
\value{
data frame of outer tree events, each made up of: time, 
        event type, recipient CompartmentType and source CompartmentType
}
\description{
Support function for sim.outer.tree.  Samples transmission, transition and
migration events based on population dynamics of the Model.
}
\keyword{internal}
