% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{.scale.contact.rates}
\alias{.scale.contact.rates}
\title{.scale.contact.rates}
\usage{
.scale.contact.rates(mx, susceptible, infected)
}
\arguments{
\item{mx:}{rate matrix for source->recipient events}

\item{size:}{vector of subpopulation sizes}
}
\value{
matrix of same dimensions as 'mx'
}
\description{
Helper function for .sample.outer.events - calculate population rates 
given per-contact rate (\beta) and population sizes of source and
recipient populations, *i.e.*, \beta*S*I
}
\keyword{internal}
