% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{.store.initial.samplings}
\alias{.store.initial.samplings}
\title{.store.initial.samplings}
\usage{
.store.initial.samplings(infected, lineages, popn.rates)
}
\arguments{
\item{infected:}{list of sampled infected Compartment objects}

\item{lineages:}{list of Lineage objects}

\item{popn.rates:}{named matrix of transmission rates (row = source)}
}
\value{
named vector of first Lineage sampling times for each Compartment
}
\description{
Stores first sampling time of a Lineage for each sampled infected 
Compartment.
}
\keyword{internal}
