% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{eventlog.from.tree}
\alias{eventlog.from.tree}
\title{eventlog.from.tree}
\usage{
eventlog.from.tree(tree)
}
\arguments{
\item{tree:}{either a Newick tree string or an object of class 'phylo' (ape)
that represents the transmission tree (history).  Internal node labels must 
be present to indicate transmission sources.}
}
\value{
An object of class EventLogger initialized with a list of fixed transmission
  events that were extracted from 'tree'.
}
\description{
\code{eventlog.from.tree} converts a Newick tree string or ape::phylo object
into a sequence of transmission events that are stored as an eventlog object
}
\seealso{
sim.outer.tree, init.branching.events
}
