% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{generate.bottleneck}
\alias{generate.bottleneck}
\title{generate.bottleneck}
\usage{
generate.bottleneck(run, comp)
}
\arguments{
\item{run}{= R6 Run object}

\item{comp}{= Compartment object}
}
\value{
lineages that 'survive' the bottleneck on towards source of transmission 
in coalescent time
}
\description{
function coalesces lineages currently extant in given Compartment to the given 
Compartment's bottleneck size bottleneck size of Compartment is user determined 
by CompartmentType
}
