% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{generate.coalescent}
\alias{generate.coalescent}
\title{generate.coalescent}
\usage{
generate.coalescent(run, line1, line2, current.time,
  is.bottleneck = FALSE)
}
\arguments{
\item{run:}{R6 Run object}

\item{line1:}{Lineage object}

\item{line2:}{Lineage object}

\item{comp:}{Compartment object}

\item{current.time:}{double; time of simulation current to this function call}
}
\value{
R6 Compartment object updated by coalescent event
}
\description{
function records the coalescent of 2 given lineages currently extant in given 
Compartment into an ancestral lineage
}
