% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{generate.migration}
\alias{generate.migration}
\title{generate.migration}
\usage{
generate.migration(run, eventlog, source, recipient, migrating.lineage,
  migration.time)
}
\arguments{
\item{run:}{MODEl object}

\item{eventlog:}{EventLogger object}

\item{migrating.lineage:}{Lineage object; to be migrated}

\item{inf:}{list of Compartment objects as potential migration sources}

\item{current.time:}{double; time of simulation current to this function call}
}
\description{
function records a migration event with a given lineage from a source to recipient migration
}
