% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDynamics.R
\name{get.counts}
\alias{get.counts}
\title{get.counts}
\usage{
get.counts(dynamics)
}
\arguments{
\item{dynamics:}{S3 object of class 'dynamics'}
}
\value{
dynamics updated with compartment sizes
}
\description{
Convert an event log into population size trajectories for every 
compartment in the model.  The user may choose to omit these counts to 
reduce the size of the eventlog `data.frame`.
}
