% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Host}
\alias{Host}
\title{Host}
\arguments{
\item{name:}{character, unique identifier for the host}

\item{compartment:}{character, reference to Compartment the host currently
belongs to.}

\item{source:}{character, reference to another Host from which a Pathogen 
lineage was transmitted to this Host.  May store more than one source
in the case of superinfection.}

\item{transmission.time:}{numeric, stores the time(s) that a Pathogen 
lineage weas transmitted to this Host.  Uses a named vector to 
make the association of time to source Host less ambiguous.}

\item{sampling.time:}{numeric, the earliest time that a Pathogen lineage 
was directly sampled from this Host.  If NA, then this is an unsampled
Host, i.e., its lineages are ancestral to lineages in subsequent 
hosts where they are sampled.}

\item{pathogens:}{list, store objects of class Pathogen associated with 
this Host}
}
\description{
\code{Host} is an R6 class representing a host individual carrying 
one or more sampled Pathogen lineages.  The transmission history among
hosts comprise the "outer" tree of the simulation.
}
