% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Host.R
\name{HostSet}
\alias{HostSet}
\title{HostSet}
\arguments{
\item{name:}{character, optional identifier}

\item{hosts:}{list, used to build a HostSet from existing Host objects}

\item{index:}{integer, internal counter}
}
\description{
A mutable set of Host objects, used to make some methods more convenient
when simulating the outer tree.
}
