% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{init.branching.events}
\alias{init.branching.events}
\title{init.branching.events}
\usage{
init.branching.events(model, eventlog)
}
\arguments{
\item{model:}{an object of class MODEL from which we will extract the 
transmission tree information.}

\item{eventlog:}{an object of class EventLog to update in-place}
}
\description{
\code{init.branching.events} initializes an EventLogger object with 
a list of fixed transmission events that are specified by the user 
in a YAML file that has been parsed into a MODEL object.
}
\seealso{
eventlog.from.tree, sim.outer.tree
}
