% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InnerTree.R
\name{InnerTree}
\alias{InnerTree}
\title{InnerTree}
\arguments{
\item{mod:}{R6 object of class Model}

\item{prefix:}{character, used for labeling Pathogen objects (default 'P')}

\item{p.index:}{integer, counter for uniquely labeling Pathogens}
}
\description{
\code{InnerTree} is an R6 class defining a persistent object that records 
events while simulating the inner tree within an outer tree.
}
\details{
The following types of events are recorded in the `inner.log`:
  sampling: `pathogen1` in `from.host`
  transmission: `from.host`->`to.host` of `pathogen1`
  coalescence: `pathogen1` and `pathogen2` within `from.host`
  bottleneck: `pathogen1` and `pathogen2` within `from.host`
}
