% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{Lineage}
\alias{Lineage}
\title{Lineage}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Lineage
}
\arguments{
\item{name:}{a character string that uniquely identifies the Lineage}

\item{type:}{a reference to an object of class LineageType (not yet implemented)}

\item{sampling.time:}{the time that the Lineage was sampled; left to NA for 
unsampled Lineages}

\item{location:}{a reference to a Compartment object}
}
\description{
\code{Lineage} is an R6 class for objects that represent pathogen lineages
that are carried by Compartments and which comprise the "inner" tree of the 
simulation.
}
\examples{
# load Compartments from a YAML object
path <- system.file('extdata', 'SI.yaml', package='twt')
settings <- yaml.load_file(path)
mod <- MODEL$new(settings)

# display first Lineage in first Compartment
comp <- mod$get.compartments()[[1]]
comp$get.lineages()  # display all 3 lineages

# manually add an unsampled Lineage
lin <- Lineage$new(name="L0", location=comp)
comp$add.lineage(lin)


}
\keyword{datasets}
