% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{load.outer.tree}
\alias{load.outer.tree}
\title{load.outer.tree}
\usage{
load.outer.tree(model)
}
\arguments{
\item{model:}{R6 object of class Model}
}
\value{
R6 object of class Run
}
\description{
`load.outer.tree`` initializes a Run object from a Model object
in which the user has explicitly listed transmission events; *i.e.*, 
bypassing simulation of the outer tree.
}
\examples{
# load model
path <- system.file('extdata', 'chain.yaml', package='twt')
settings <- yaml.load_file(path)
mod <- Model$new(settings)

run <- load.outer.tree(mod)
plot(run$get.eventlog())

}
\seealso{
attach.tree, sim.outer.tree
}
