% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{Model}
\alias{Model}
\title{Model}
\arguments{
\item{settings:}{a named list returned by `yaml.load_file()` that contains 
user specifications of the simulation model.}
}
\description{
\code{Model} is an R6 class that defines an object that parses the model 
settings from a YAML (Parameters, Compartments, Sampling).
}
\details{
Four types of events affect Compartments:
  births         0->S      dS/dt = a [or] aS
  deaths         S->0      dS/dt = -a [or] aS
  migrations     I->S      dI/dt = -aI, dS/dt = aI
  transmissions  S+I->I+I  dS/dt = -bSI, dI/dt = bSI
In the case of superinfection, transmission occurs within a Compartment
  I + I -> I + I
This does not affect Compartment size and dynamics, but it is a branching
event.

It is important that individual model terms are separated!  For example,
there may be birth and death rates associated with a Compartment:
  dS/dt = lambda S - delta S [YES]
        = (lambda-delta)S    [NO!]
In other words, we want to model these stochastic events separately 
instead of the net rate of change.
}
\section{Fields}{

\describe{
\item{\code{parameters}}{model parameters}

\item{\code{compartments}}{vector of Compartment objects}

\item{\code{sampling}}{sampling conditions}
}}

\examples{
require(R6)
require(yaml)
require(igraph)
settings <- yaml.load_file("examples/SIRS_serial.yaml")
mod <- Model$new(settings)
summary(mod)
plot(mod)

}
