% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OuterTree.R
\name{OuterTree}
\alias{OuterTree}
\title{OuterTree}
\arguments{
\item{mod:}{R6 object of class Model}
}
\description{
\code{OuterTree} is an R6 class defining a persistent object that records 
events while simulating the outer tree from population trajectories.
}
