% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathogen.R
\name{Pathogen}
\alias{Pathogen}
\title{Pathogen}
\arguments{
\item{name:}{character, unique identifier for Pathogen object}

\item{start.time:}{numeric, start time of lineage; a coalescent event 
terminates the parent and starts two child lineages.}

\item{end.time:}{numeric, end time of lineage; associated with a 
coalescent event (if parent) or sampling event.}

\item{parent:}{character, parent Pathogen object}

\item{children:}{list, child Pathogen objects (NOTE: avoid cloning 
Pathogen objects or there will be a circular reference problem!)}
}
\description{
\code{Pathogen} is an R6 class for objects that represent pathogen lineages
that are carried by Hosts and which comprise the "inner" tree of the 
simulation.
}
