% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventLogger.R
\name{plot.EventLogger}
\alias{plot.EventLogger}
\title{plot.EventLogger}
\usage{
\method{plot}{EventLogger}(eventlog, transmissions = FALSE,
  migrations = FALSE, node.labels = FALSE)
}
\arguments{
\item{eventlog:}{Object of class 'EventLogger'}

\item{transmissions:}{If TRUE, display transmission events in plot.}
}
\description{
Generate a plot summarizing the content of an EventLogger object
as a phylogenetic tree.
}
