% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{plot.Model}
\alias{plot.Model}
\title{plot.Model
Plot a graph summarizing compartments and rates
TODO: label edges with rate expressions}
\usage{
\method{plot}{Model}(obj)
}
\description{
plot.Model
Plot a graph summarizing compartments and rates
TODO: label edges with rate expressions
}
