% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run.R
\name{plot.Run}
\alias{plot.Run}
\title{Wrapper around plot.EventLogger}
\usage{
\method{plot}{Run}(run, transmissions = FALSE, migrations = FALSE, node.labels = FALSE)
}
\description{
Wrapper around plot.EventLogger
}
