% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDynamics.R
\name{plot.twt.counts}
\alias{plot.twt.counts}
\title{plot.twt.counts}
\usage{
\method{plot}{twt.counts}(counts, pal = NA, xlab = "Time", ylab = "Frequency", lwd = 2, ...)
}
\arguments{
\item{counts:}{S3 object of class twt.counts}

\item{pal:}{character, palette for drawing lines; defaults to Dark2}

\item{xlab:}{character, label for x-axis (default: 'Time')}

\item{ylab:}{character, label for y-axis (default: 'Frequency')}

\item{lwd:}{numeric, line width for plotting}

\item{...:}{additional arguments passed to base plot() function}
}
\description{
S3 method for visualizing the simulated population dynamics from 
simulate.dynamics().
}
