% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{print.Model}
\alias{print.Model}
\title{print.Model
S3 class function to display contents of a Model object}
\usage{
\method{print}{Model}(obj)
}
\description{
print.Model
S3 class function to display contents of a Model object
}
