% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{remove.lineage.pairs}
\alias{remove.lineage.pairs}
\title{remove.lineage.pairs}
\usage{
remove.lineage.pairs(model, lineage)
}
\arguments{
\item{model}{= MODEL object, one per simulation}

\item{lineage}{= Lineage object, to be removed from list of lineage pair choices}
}
\description{
this function removes lineage pairs in MODEL obj attr `choices`: list of lineage 
pair choices to coalesce
}
