% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{resolve.bottleneck}
\alias{resolve.bottleneck}
\title{resolve.bottleneck}
\usage{
resolve.bottleneck(run, comp)
}
\arguments{
\item{run:}{R6 object of class Run}

\item{comp:}{R6 object of class Compartment}
}
\value{
list of Lineage objects to pass to source Compartment
}
\description{
Helper function records the bottleneck of extant Lineages, with a
random sample being transferred to the source Compartment.
}
