% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{resolve.coalescent}
\alias{resolve.coalescent}
\title{resolve.coalescent}
\usage{
resolve.coalescent(run, comp, time, is.bottleneck = FALSE)
}
\arguments{
\item{run:}{an R6 object of class Run}

\item{comp:}{an R6 object of class Compartment}

\item{time:}{double; time of coalescent event}

\item{is.bottleneck:}{if TRUE, mark event as 'bottleneck' in log}
}
\description{
Helper function records the coalescence of two extant Lineages into an
ancestral Lineage.
}
