% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{resolve.migration}
\alias{resolve.migration}
\title{resolve.migration}
\usage{
resolve.migration(run, inf, migration.event)
}
\arguments{
\item{run:}{Run object}

\item{inf:}{list of sampled and unsampled infected compartments 
of type Compartment}

\item{migration.event:}{list of information specific to the migration 
event being resolved}
}
\description{
function resolves a potential migration event generated from 
`sim.migrations()`
}
