% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{resolve.transition}
\alias{resolve.transition}
\title{resolve.transition}
\usage{
resolve.transition(run, e)
}
\arguments{
\item{run:}{R6 object of class Run}

\item{e:}{a row from an EventLogger dataframe}
}
\description{
Helper function records the transition of a Compartment from one Type
to another.  Note this reverts a Compartment at time 0 (most recent time)
from derived type2 to ancestral type1.
}
