% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run.R
\docType{data}
\name{Run}
\alias{Run}
\title{Run}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Run
}
\arguments{
\item{model:}{an object of class Model}
}
\description{
\code{Run} is an R6 class that is derived from \code{Model}.  It inherits
all simulation model features including CompartmentTypes, Compartments, 
and Lineages.  Unlike \code{Model}, a \code{Run} object is used to track 
an outcome of the simulation model and is therefore mutable where \code{Model}
is immutable.  For example, \code{Run} tracks which lineages are extant and 
the numbers of pairs of Lineages that can coalesce over time.  In addition, 
the number of unsampled compartments is stochastic outcome of the model that
is tracked at this level.
}
\section{Fields}{

\describe{
\item{\code{extant.lineages}}{list of Lineage objects with sampling time t=0 (most recent)}

\item{\code{locations}}{a named list of Lineage objects keyed by Compartment}

\item{\code{counts}}{a data frame of population dynamics}
}}

\keyword{datasets}
