% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{sample.coalescents}
\alias{sample.coalescents}
\title{sample.coalescents}
\usage{
sample.coalescents(run, current.time)
}
\arguments{
\item{run:}{an R6 object of class Run'}

\item{current.time:}{current simulation time for inner tree}
}
\value{
Named numeric vector of waiting times per extant compartment
}
\description{
Draw waiting times for all compartments that have two or more extant lineages.
}
\examples{
# this model specifies 10 compartments with 3 lineages each,
# and constant coalescent rate 1.0
path <- system.file('extdata', 'SI.yaml', package='twt')
settings <- yaml.load_file(path)
mod <- MODEL$new(settings)
run <- Run$new(mod)

# these should average 0.333
sample.coalescents(run, 0)

}
