% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{sim.inner.tree}
\alias{sim.inner.tree}
\title{sim.inner.tree}
\usage{
sim.inner.tree(outer, mod)
}
\arguments{
\item{outer:}{R6 object of class `OuterTree`}

\item{mod:}{R6 object of class `Model`}
}
\value{
R6 object of class `InnerTree`
}
\description{
Simulate the coalescence of Pathogen lineages within Hosts, and resolve 
superinfection events that may involve sampled lineages.
}
