% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMigrations.R
\name{sim.migrations}
\alias{sim.migrations}
\title{sim.migrations}
\usage{
sim.migrations(model, eventlog)
}
\arguments{
\item{model:}{R6 object from MODEL$new()}

\item{eventlog:}{R6 Eventlogger object populated by sim.outer.tree()}
}
\description{
Annotate and update the EventLogger object with migration events.
}
