% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{summary.Model}
\alias{summary.Model}
\title{summary.Model
S3 class function to summarize a Model object - simply a wrapper 
around print()}
\usage{
\method{summary}{Model}(obj)
}
\description{
summary.Model
S3 class function to summarize a Model object - simply a wrapper 
around print()
}
