% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{update.transmission}
\alias{update.transmission}
\title{update.transmission}
\usage{
\method{update}{transmission}(model, eventlog, inf, inf.names,
  transm.event)
}
\arguments{
\item{model:}{MODEL object}

\item{eventlog:}{EventLogger object}

\item{inf:}{list of sampled and unsampled infected compartments of type Compartment}

\item{inf.names:}{list of sampled and unsampled infected compartments names of type character}

\item{transm.event:}{list of information specific to the transmission event being resolved}
}
\description{
\code{update.transmission} resolves lineages associated with a transmission event 
between compartments and updates the eventlog.
}
