% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{wait.time}
\alias{wait.time}
\title{wait.time}
\usage{
wait.time(k, t1, alpha, beta)
}
\arguments{
\item{k:}{number of extant lineages that could potentially coalesce}

\item{t1:}{start time in reverse, relative to intercept}

\item{alpha:}{intercept, population size at t=0}

\item{beta:}{slope of linear growth interval}
}
\value{
Random variate from waiting time distribution.
}
\description{
Draw waiting time to next coalescent event for the current interval
of a piecewise linear model describing the population 
growth dynamics over time.
Inverse cumulative function from Romero-Severson et al., 2017 
https://doi.org/10.1534/genetics.117.300284
}
\keyword{internal}
