% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventLogger.R
\name{as.phylo.EventLogger}
\alias{as.phylo.EventLogger}
\title{as.phylo}
\usage{
\method{as.phylo}{EventLogger}(eventlog, transmissions = FALSE,
  migrations = FALSE, node.labels = FALSE)
}
\arguments{
\item{eventlog:}{EventLogger object}

\item{transmissions:}{logical; if TRUE, transmission events included, 
else excluded otherwise.}
}
\value{
ape::phylo object
}
\description{
An S3 method converts events stored in the EventLogger object into an inner 
coalescent tree w/ option to include/exclude transmission events.
}
