% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{calc.coal.wait.times}
\alias{calc.coal.wait.times}
\title{calc.coal.wait.times}
\usage{
calc.coal.wait.times(run, current.time, dynamic = FALSE)
}
\arguments{
\item{run:}{an R6 object of class Run'}

\item{current.time:}{current simulation time for inner tree}
}
\value{
Named numeric vector of waiting times per compartment
}
\description{
Draw waiting times for all compartments that have two or more extant lineages.
}
\examples{
# this model specifies 10 compartments with 3 lineages each,
# and constant coalescent rate 1.0
path <- system.file('extdata', 'SI.yaml', package='twt')
settings <- yaml.load_file(path)
mod <- MODEL$new(settings)
run <- Run$new(mod)

# these should average 0.333
calc.coal.wait.times(run, 0)

}
