% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Compartment}
\alias{Compartment}
\title{Compartment}
\description{
\code{Compartment} is an R6 class for objects that represent the units of an
outer tree simulation, such as a host individual or deme.
}
\examples{
# load Compartments from a YAML object
path <- system.file('extdata', 'SI.yaml', package='twt')
settings <- yaml.load_file(path)
mod <- MODEL$new(settings)

# display first Compartment object
host1 <- mod$get.compartments()[[1]]
host1

# manually initialize a new Compartment object
hostN <- Compartment$new(name='newHost', unsampled=TRUE)
hostN$set.type(host1$get.type())

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Compartment$new()}}
\item \href{#method-copy}{\code{Compartment$copy()}}
\item \href{#method-get.name}{\code{Compartment$get.name()}}
\item \href{#method-get.type}{\code{Compartment$get.type()}}
\item \href{#method-get.source}{\code{Compartment$get.source()}}
\item \href{#method-get.branching.time}{\code{Compartment$get.branching.time()}}
\item \href{#method-set.type}{\code{Compartment$set.type()}}
\item \href{#method-set.source}{\code{Compartment$set.source()}}
\item \href{#method-set.branching.time}{\code{Compartment$set.branching.time()}}
\item \href{#method-is.unsampled}{\code{Compartment$is.unsampled()}}
\item \href{#method-get.lineages}{\code{Compartment$get.lineages()}}
\item \href{#method-add.lineage}{\code{Compartment$add.lineage()}}
\item \href{#method-remove.lineage}{\code{Compartment$remove.lineage()}}
\item \href{#method-clone}{\code{Compartment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartment$new(
  name = NA,
  type = NA,
  source = NA,
  branching.time = NA,
  unsampled = FALSE,
  lineages = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copy"></a>}}
\subsection{Method \code{copy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartment$copy(deep = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.name"></a>}}
\subsection{Method \code{get.name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartment$get.name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.type"></a>}}
\subsection{Method \code{get.type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartment$get.type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.source"></a>}}
\subsection{Method \code{get.source()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartment$get.source()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.branching.time"></a>}}
\subsection{Method \code{get.branching.time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartment$get.branching.time()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set.type"></a>}}
\subsection{Method \code{set.type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartment$set.type(new.type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set.source"></a>}}
\subsection{Method \code{set.source()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartment$set.source(new.source)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set.branching.time"></a>}}
\subsection{Method \code{set.branching.time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartment$set.branching.time(new.branching.time)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is.unsampled"></a>}}
\subsection{Method \code{is.unsampled()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartment$is.unsampled()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.lineages"></a>}}
\subsection{Method \code{get.lineages()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartment$get.lineages()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add.lineage"></a>}}
\subsection{Method \code{add.lineage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartment$add.lineage(new.lineage)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove.lineage"></a>}}
\subsection{Method \code{remove.lineage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartment$remove.lineage(ex.lineage)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
