% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{CompartmentType}
\alias{CompartmentType}
\title{CompartmentType}
\description{
\code{CompartmentType} is an R6 class that defines a type of compartment, 
such as a class of host individual (risk group) with a specific transmission 
rate.
}
\examples{

# load CompartmentTypes from a YAML object
path <- system.file('extdata', 'SI.yaml', package='twt')
settings <- yaml.load_file(path)
mod <- Model$new(settings)
mod$get.types()

# manually specify a CompartmentType object (usually done by YAML)
host <- CompartmentType$new(name='host', branching.rates=c(host=0.1), 
bottleneck.size=1, coalescent.rate=1)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CompartmentType$new()}}
\item \href{#method-get.bottleneck.size}{\code{CompartmentType$get.bottleneck.size()}}
\item \href{#method-get.name}{\code{CompartmentType$get.name()}}
\item \href{#method-get.unsampled}{\code{CompartmentType$get.unsampled()}}
\item \href{#method-get.branching.rates}{\code{CompartmentType$get.branching.rates()}}
\item \href{#method-get.branching.rate}{\code{CompartmentType$get.branching.rate()}}
\item \href{#method-get.migration.rates}{\code{CompartmentType$get.migration.rates()}}
\item \href{#method-get.migration.rate}{\code{CompartmentType$get.migration.rate()}}
\item \href{#method-get.transition.rates}{\code{CompartmentType$get.transition.rates()}}
\item \href{#method-get.transition.rate}{\code{CompartmentType$get.transition.rate()}}
\item \href{#method-get.coalescent.rate}{\code{CompartmentType$get.coalescent.rate()}}
\item \href{#method-get.popn.growth.dynamics}{\code{CompartmentType$get.popn.growth.dynamics()}}
\item \href{#method-get.transmission.times}{\code{CompartmentType$get.transmission.times()}}
\item \href{#method-set.transmission.times}{\code{CompartmentType$set.transmission.times()}}
\item \href{#method-set.migration.rate}{\code{CompartmentType$set.migration.rate()}}
\item \href{#method-clone}{\code{CompartmentType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentType$new(
  name = NA,
  unsampled = NA,
  branching.rates = NA,
  transition.rates = NA,
  migration.rates = NA,
  bottleneck.size = NA,
  coalescent.rate = NA,
  popn.growth.dynamics = NA,
  transmission.times = NA
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.bottleneck.size"></a>}}
\subsection{Method \code{get.bottleneck.size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentType$get.bottleneck.size()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.name"></a>}}
\subsection{Method \code{get.name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentType$get.name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.unsampled"></a>}}
\subsection{Method \code{get.unsampled()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentType$get.unsampled()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.branching.rates"></a>}}
\subsection{Method \code{get.branching.rates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentType$get.branching.rates()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.branching.rate"></a>}}
\subsection{Method \code{get.branching.rate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentType$get.branching.rate(name.type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.migration.rates"></a>}}
\subsection{Method \code{get.migration.rates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentType$get.migration.rates()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.migration.rate"></a>}}
\subsection{Method \code{get.migration.rate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentType$get.migration.rate(name.type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.transition.rates"></a>}}
\subsection{Method \code{get.transition.rates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentType$get.transition.rates()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.transition.rate"></a>}}
\subsection{Method \code{get.transition.rate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentType$get.transition.rate(name.type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.coalescent.rate"></a>}}
\subsection{Method \code{get.coalescent.rate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentType$get.coalescent.rate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.popn.growth.dynamics"></a>}}
\subsection{Method \code{get.popn.growth.dynamics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentType$get.popn.growth.dynamics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.transmission.times"></a>}}
\subsection{Method \code{get.transmission.times()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentType$get.transmission.times()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set.transmission.times"></a>}}
\subsection{Method \code{set.transmission.times()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentType$set.transmission.times(vector.transm.times)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set.migration.rate"></a>}}
\subsection{Method \code{set.migration.rate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentType$set.migration.rate(recipient.type, new.migr.rate)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
