% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{CompartmentType}
\alias{CompartmentType}
\title{CompartmentType}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CompartmentType
}
\arguments{
\item{name:}{a character string that uniquely identifies the class}

\item{unsampled:}{if TRUE, no Compartments of this Type can contain sampled
lineages (directly observed tips of the inner tree).}

\item{susceptible:}{if TRUE, no Compartments of this Type contain either
sampled or unsampled lineages.}

\item{branching.rates:}{a named vector of transmission rates *to* other
CompartmentTypes.}

\item{migration.rates:}{a named vector of migration rates *to* other 
CompartmentTypes.}

\item{bottleneck.size:}{the maximum number of lineages that can be transmitted
to a Compartment of this Type.}

\item{coalescent.rate:}{the rate at which lineages coalesce within a Compartment
of this Type.}

\item{death.rate.distr:}{a text expression for the waiting time distribution
to a death event. Required for `popn.growth.dynamics` to override 
`coalescent.rate`.}

\item{wait.time.distr:}{a text expression for the waiting time distribution
to a transmission event.}

\item{popn.growth.dynamics:}{a text expression for population growth dynamics
in forward time. If not NULL, can override `coalescent.rate`.}

\item{transmission.times:}{numeric vector of transmission event times, 
populated by `outer.tree.sim` from class parameters.}
}
\description{
\code{CompartmentType} is an R6 class that defines a type of compartment, 
such as a class of host individual (risk group) with a specific transmission 
rate.
}
\keyword{datasets}
