% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{.assign.events}
\alias{.assign.events}
\title{.assign.events}
\usage{
.assign.events(run, events)
}
\description{
Assignment of outer events proceeds in reverse (coalescent) time,
starting from the sampled Compartments.
TODO: access population dynamics through Run$get.counts() instead of 
      cbind'ed events
}
