% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMigrations.R
\name{.calc.migration.events}
\alias{.calc.migration.events}
\title{.calc.migration.events}
\usage{
.calc.migration.events(popn.totals, popn.migration.rates,
  max.sampling.time, possible.source.types, transmission.events)
}
\arguments{
\item{popn.totals:}{totals at time `t=0` of susceptible and infected specific 
to each CompartmentType}

\item{popn.migration.rates:}{rates of migration between different 
CompartmentTypes}

\item{max.sampling.time:}{time of the most recent sample as a endpoint for 
this function (type numeric)}

\item{possible.source.types:}{list of possible Sources that each recipient 
type can receive a migrating lineage from}
}
\value{
m_events = data frame of migration events, each made up of: time, 
recipient Type, and source Type
}
\description{
Generates migration events only, based on population dynamics of the MODEL.
}
\keyword{internal}
