% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{.calc.transmission.events}
\alias{.calc.transmission.events}
\title{.calc.transmission.events}
\usage{
.calc.transmission.events(popn.totals, popn.rates, init.samplings,
  possible.sources)
}
\arguments{
\item{popn.totals:}{totals at time `t=0` of susceptible and infected specific to 
each CompartmentType}

\item{popn.rates:}{rates of transmission between different CompartmentTypes}

\item{init.samplings:}{list of first sampling times for each Compartment}

\item{possible.sources:}{list of possible Sources that each recipient Type can 
receive a transmission from}
}
\value{
t_events: data frame of transmission events, each made up of: time, 
        recipient Type, and source Type
}
\description{
Generates transmission events only, based on population dynamics of the MODEL
}
\keyword{internal}
