% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{.calc.transmission.events}
\alias{.calc.transmission.events}
\title{.calc.transmission.events}
\usage{
.calc.transmission.events(init.conds, popn.rates, init.samplings,
  max.attempts = 10)
}
\arguments{
\item{init.conds:}{initial conditions, comprising (1) time scale of simulation
(2) initial population size per CompartmentType and (3) Type
of index case}

\item{popn.rates:}{rates of transmission between different CompartmentTypes}

\item{init.samplings:}{list of first sampling times for each Compartment}

\item{max.attempts:}{number of tries to sample transmission events}
}
\value{
data frame of transmission events, each made up of: time, 
        recipient Type, source Type, and LineageType (not yet implemented)
}
\description{
Generates transmission events only, based on population dynamics of the
MODEL
}
\keyword{internal}
