% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run.R
\name{.plot.outer.tree}
\alias{.plot.outer.tree}
\title{.plot.outer.tree}
\usage{
.plot.outer.tree(run, segments.lwd = 5, ...)
}
\arguments{
\item{e:}{EventLogger object}

\item{segments.lwd:}{line width for segments}
}
\description{
A function that converts events stored in an EventLogger object into
an outer transmission tree.
}
\examples{
path <- system.file('extdata', 'structSI.yaml', package='twt')
settings <- yaml.load_file(path)
model <- Model$new(settings)
run <- sim.outer.tree(model)
plot(run)

}
\keyword{internal}
