% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{.sample.outer.events}
\alias{.sample.outer.events}
\title{.sample.outer.events}
\usage{
.sample.outer.events(run, max.attempts = 10)
}
\arguments{
\item{run:}{R6 object of class Run}

\item{max.attempts:}{number of tries to sample events}
}
\value{
data frame of outer tree events, each made up of: time, 
        event type, recipient CompartmentType and source CompartmentType
}
\description{
Support function for sim.outer.tree.  Samples transmission, transition and
migration events based on population dynamics of the Model.
}
\keyword{internal}
