% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{.store.initial.samplings}
\alias{.store.initial.samplings}
\title{.store.initial.samplings}
\usage{
.store.initial.samplings(infected, types, lineages)
}
\arguments{
\item{infected}{= list of infected Compartment objects}

\item{types}{= list of CompartmentType objects}

\item{lineages}{= list of Lineage objects}
}
\value{
named list of possible Types of `source` with recipient types as names
and list of first Lineage sampling times for each Compartment
}
\description{
Stores first sampling time of a Lineage for each sampled infected 
Compartment.
}
\keyword{internal}
