% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventLogger.R
\docType{data}
\name{EventLogger}
\alias{EventLogger}
\title{EventLogger}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EventLogger
}
\arguments{
\item{events:}{a data frame where each row represents an event with a time 
stamp in forward time.}

\item{type:}{event type, one of 'transmission', 'transition', 'migration', 
'coalescence' or 'bottleneck'.}

\item{time:}{CUMULATIVE time that event has occurred between two compartments 
       in a transmission/migration/coalescent event
Record which Lineages are transmitted from source to 
recipient (should only be one entry with Compartment as recipient)}

\item{recipient:}{Compartment object}

\item{lineages:}{a vector of names of Lineages to transfer out
                  of recipient Compartment
Record which Lineages are transmitted from source to recipient
through a migration event}

\item{recipient:}{Compartment object}

\item{source:}{Compartment object}

\item{time:}{double, time of migration event}

\item{lineages:}{a list of Lineage objects}
}
\description{
\code{EventLogger} is an R6 class for an object that tracks migration, 
transmission, and coalescent events.  Note that bottleneck events are logged as 
coalescent events.
}
\examples{
# manually initialize an EventLog object
e <- EventLogger$new()
# note this log entry is not linked to existing Lineage or Compartment objects
e$add.event("transmission", time=1, line1="NA", comp1="hist1", comp2="host2")
e$get.all.events()

}
\keyword{datasets}
