% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{EventLogger}
\alias{EventLogger}
\title{EventLogger}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EventLogger
}
\arguments{
\item{events:}{a data frame where each row represents an event with a time 
stamp in forward time.}

\item{events.noncumul:}{optionally, a data frame where events are stamped 
with the waiting time since the preceding event}

\item{migration.events.storage:}{a data frame of migration events, returned by
`simMigrations.R:.calc.migration.events()`.}
}
\description{
\code{EventLogger} is an R6 class for an object that tracks migration, 
transmission, and coalescent events.  Note that bottleneck events are logged as 
coalescent events.
}
\keyword{datasets}
