% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventLogger.R
\name{EventLogger}
\alias{EventLogger}
\title{EventLogger}
\description{
\code{EventLogger} is an R6 class for an object that tracks migration, 
transmission, and coalescent events.  Note that bottleneck events are logged as 
coalescent events.
}
\examples{
# manually initialize an EventLog object
e <- EventLogger$new()
# note this log entry is not linked to existing Lineage or Compartment objects
e$add.event("transmission", time=1, line1="NA", comp1="hist1", comp2="host2")
e$get.all.events()

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EventLogger$new()}}
\item \href{#method-get.all.events}{\code{EventLogger$get.all.events()}}
\item \href{#method-get.events}{\code{EventLogger$get.events()}}
\item \href{#method-add.event}{\code{EventLogger$add.event()}}
\item \href{#method-clear.events}{\code{EventLogger$clear.events()}}
\item \href{#method-modify.event}{\code{EventLogger$modify.event()}}
\item \href{#method-get.migration.events}{\code{EventLogger$get.migration.events()}}
\item \href{#method-store.migration.events}{\code{EventLogger$store.migration.events()}}
\item \href{#method-get.fixed.samplings}{\code{EventLogger$get.fixed.samplings()}}
\item \href{#method-store.fixed.samplings}{\code{EventLogger$store.fixed.samplings()}}
\item \href{#method-clone}{\code{EventLogger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventLogger$new(events = NA, migration.events.storage = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.all.events"></a>}}
\subsection{Method \code{get.all.events()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventLogger$get.all.events()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.events"></a>}}
\subsection{Method \code{get.events()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventLogger$get.events(event.type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add.event"></a>}}
\subsection{Method \code{add.event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventLogger$add.event(
  type,
  time,
  line1 = NA,
  line2 = NA,
  comp1 = NA,
  comp2 = NA,
  type1 = NA,
  type2 = NA
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear.events"></a>}}
\subsection{Method \code{clear.events()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventLogger$clear.events()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-modify.event"></a>}}
\subsection{Method \code{modify.event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventLogger$modify.event(transmission.time, lineages)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.migration.events"></a>}}
\subsection{Method \code{get.migration.events()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventLogger$get.migration.events()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-store.migration.events"></a>}}
\subsection{Method \code{store.migration.events()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventLogger$store.migration.events(migration.events)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.fixed.samplings"></a>}}
\subsection{Method \code{get.fixed.samplings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventLogger$get.fixed.samplings()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-store.fixed.samplings"></a>}}
\subsection{Method \code{store.fixed.samplings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventLogger$store.fixed.samplings(model.fixed.samplings)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventLogger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
