% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{generate.coalescent}
\alias{generate.coalescent}
\title{generate.coalescent}
\usage{
generate.coalescent(model, eventlog, lineages.to.coalesce, coalescing.comp,
  current.time)
}
\arguments{
\item{model}{= MODEL object}

\item{eventlog}{= EventLogger object}

\item{lineages.to.coalesce}{= vector of Lineage objects of length 2}

\item{coalescing.comp}{= Compartment object}

\item{current.time}{= double; time of simulation current to this function call}
}
\description{
function records the coalescent of 2 given lineages currently extant in given 
Compartment into an ancestral lineage
}
