% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{generate.migration}
\alias{generate.migration}
\title{generate.migration}
\usage{
generate.migration(run, source, recipient, migrating.lineage,
  migration.time)
}
\arguments{
\item{run:}{Run object}

\item{migrating.lineage:}{Lineage object; to be migrated}

\item{inf:}{list of Compartment objects as potential migration sources}

\item{current.time:}{double; time of simulation current to this function call}
}
\description{
Function to record a migration event of a given Lineage from a source to 
recipient compartment.
}
