% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDynamics.R
\name{get.counts}
\alias{get.counts}
\title{get.counts}
\usage{
get.counts(eventlog, mod, chunk.size = 100)
}
\arguments{
\item{eventlog:}{data.frame or character, result from sim.dynamics()}

\item{mod:}{R6 object of class Model}

\item{chunk.size:}{integer, used to allocate new rows for output data frame}
}
\value{
data.frame of compartment sizes over time
}
\description{
Convert an event log into population size trajectories for every 
compartment in the model.  We omit these to keep the event log compact.
}
\examples{
counts <- get.counts("eventlog.csv", mod)
plot(counts)  # S3 method

}
