% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDynamics.R
\name{get.counts}
\alias{get.counts}
\title{get.counts}
\usage{
get.counts(eventlog, mod, chunk.size = 100)
}
\description{
Convert an event log into population size trajectories for every 
compartment in the model.  We omit these to keep the event log compact.
}
\examples{
counts <- get.counts("eventlog.csv", mod)
plot(counts)  # S3 method

}
