% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadInputs.R
\docType{data}
\name{MODEL}
\alias{MODEL}
\title{MODEL}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MODEL
}
\arguments{
\item{settings:}{a named list returned by `yaml.load_file()` that contains 
user specifications of the simulation model.}
}
\description{
\code{MODEL} is an R6 class that defines an object that generates all of the 
objects of the various classes (e.g., Compartment, CompartmentType) that 
define a simulation model.
}
\section{Fields}{

\describe{
\item{\code{origin.times}}{matrix with rows for every LineageType, and columns for 
the origin time (in reverse, relative to most recent lineage sampling time t=0) 
followed by number of susceptibles per CompartmentType at origin.}

\item{\code{types}}{vector of CompartmentType objects}

\item{\code{compartments}}{vector of Compartment objects}

\item{\code{lineages}}{vector of Lineage objects}

\item{\code{extant.lineages}}{list of Lineage objects with sampling time t=0 (most recent)}

\item{\code{locations}}{a named list of Lineage objects keyed by Compartment}

\item{\code{choices}}{list of Lineage pairs that may coalesce per Compartment, 
keyed by index to `locations`}

\item{\code{fixed.samplings}}{list of Lineage names and sampling times for plotting}
}}

\examples{
require(twt)
path <- system.file('extdata', 'SI.yaml', package='twt')
settings <- yaml.load_file(path)
mod <- MODEL$new(settings)

}
\keyword{datasets}
