% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\docType{data}
\name{Model}
\alias{Model}
\title{Model}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Model
}
\arguments{
\item{settings:}{a named list returned by `yaml.load_file()` that contains 
user specifications of the simulation model.}
}
\description{
\code{Model} is an R6 class that defines an object that generates all of the 
objects of the various classes (e.g., Compartment, CompartmentType) that 
define a simulation model.  A \code{Model} object is immutable - it should
not change over the course of simulation.  Instead, we derive a \code{Run}
object class that inherits from a \code{Model}.
}
\section{Fields}{

\describe{
\item{\code{initial.conds}}{Initial Conditions}

\item{\code{types}}{vector of CompartmentType objects}

\item{\code{compartments}}{vector of Compartment objects}

\item{\code{lineages}}{vector of Lineage objects}

\item{\code{fixed.samplings}}{list of Lineage names and sampling times for plotting}
}}

\examples{
require(twt)
# get path to example YAML file
path <- system.file('extdata', 'SI.yaml', package='twt')

# load file and parse to construct Model object
settings <- yaml.load_file(path)
mod <- Model$new(settings)

# display summary information (calls S3 print method)
mod

}
\keyword{datasets}
