% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Pathogen}
\alias{Pathogen}
\title{Pathogen}
\arguments{
\item{name:}{character, unique identifier for Pathogen object}

\item{sampling.time:}{numeric, time that the Pathogen was sampled; left to 
NA for unsampled Pathogens}

\item{host:}{a reference to a Host object}
}
\description{
\code{Pathogen} is an R6 class for objects that represent pathogen lineages
that are carried by Hosts and which comprise the "inner" tree of the 
simulation.
}
