% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventLogger.R
\name{plot.EventLogger}
\alias{plot.EventLogger}
\title{plot.EventLogger}
\usage{
\method{plot}{EventLogger}(eventlog, transmissions = FALSE, migrations = FALSE, node.labels = FALSE)
}
\arguments{
\item{eventlog:}{Object of class 'EventLogger'}

\item{transmissions:}{If TRUE, display transmission events in plot.}
}
\description{
Generate a plot summarizing the content of an EventLogger object
as a phylogenetic tree.
}
\examples{
# load model
path <- system.file('extdata', 'SI.yaml', package='twt')

# load file and parse to construct MODEL object
settings <- yaml.load_file(path)
mod <- MODEL$new(settings)

# simulate outer tree
outer <- sim.outer.tree(mod)

# simulate inner tree
tree <- sim.inner.tree(mod, outer)
plot(tree)

}
