% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventLogger.R
\name{plot.EventLogger}
\alias{plot.EventLogger}
\title{plot.EventLogger
Generic plot method for EventLogger class as an S3 object.
Converts the EventLogger object into an ape:phylo S3 object.}
\usage{
\method{plot}{EventLogger}(eventlog, transmissions, migrations)
}
\arguments{
\item{eventlog:}{R6 object of class EventLogger}

\item{transmissions:}{if TRUE, display transmission events on inner tree
plot.}

\item{migrations:}{if TRUE, display migration events on inner tree plot.}
}
\description{
plot.EventLogger
Generic plot method for EventLogger class as an S3 object.
Converts the EventLogger object into an ape:phylo S3 object.
}
\examples{
path <- system.file('extdata', 'structSI.yaml', package='twt')
settings <- yaml.load_file(path)
structSI <- Model$new(settings)
run <- sim.outer.tree(structSI)
eventlog <- sim.inner.tree(run)
tr <- as.phylo(eventlog)
plot(tr)

}
