% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OuterTree.R
\name{plot.OuterTree}
\alias{plot.OuterTree}
\title{plot.OuterTree
S3 generic plot method for objects of class `outerTree`.  This visualizes
the collection of events sampled in the outer simulation as a transmission 
tree.}
\usage{
\method{plot}{OuterTree}(obj, pad = 1.05)
}
\description{
plot.OuterTree
S3 generic plot method for objects of class `outerTree`.  This visualizes
the collection of events sampled in the outer simulation as a transmission 
tree.
}
