% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run.R
\name{plot.Run}
\alias{plot.Run}
\title{plot.Run}
\usage{
\method{plot}{Run}(run, type = "t", ...)
}
\arguments{
\item{run:}{R6 object of class `Run`}

\item{type:}{what type of plot should be drawn.  Possible types are
\itemize{
\item "t" for transmission tree plot
\item "s" for stair step plot of population dynamics
}}
}
\description{
S3 plot method for Run objects.
}
\details{
If the Run object is the outcome of sim.outer.tree() or either of the 
related outer tree functions, then we call the internal function 
.plot.outer.tree().

If the Run object is the outcome of sim.inner.tree(), then the eventlog 
is converted into a Phylo object and plotted using the S3 method provided 
by the `ape` package.
}
\examples{
# load model
path <- system.file('extdata', 'SI.yaml', package='twt')

# load file and parse to construct MODEL object
settings <- yaml.load_file(path)
mod <- Model$new(settings)

# simulate outer tree
run <- sim.outer.tree(mod)

# generate plots side-by-side
par(mfrow=c(1,2))
plot(run, type='s')  # population dynamics
plot(run, type='t')  # transmission tree

}
