% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventLogger.R
\name{print.EventLogger}
\alias{print.EventLogger}
\title{print.EventLogger}
\usage{
\method{print}{EventLogger}(eventlog)
}
\arguments{
\item{eventlog:}{object of class 'EventLog'}
}
\description{
`print.EventLogger` is simply a wrapper function on EventLogger's 
`get.all.events()`.
}
\examples{
e <- EventLogger$new()
e  # no events to display

e$add.event("transmission", time=1, line1="NA", comp1="hist1", comp2="host2")
e

}
