% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventLogger.R
\name{print.EventLogger}
\alias{print.EventLogger}
\title{print.EventLogger}
\usage{
\method{print}{EventLogger}(eventlog)
}
\arguments{
\item{eventlog:}{object of class 'EventLog'}
}
\description{
`print.EventLogger` is simply a wrapper function on EventLogger's 
`get.all.events()`.
}
