% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{remove.lineage.pairs}
\alias{remove.lineage.pairs}
\title{remove.lineage.pairs}
\usage{
remove.lineage.pairs(run, lineage)
}
\arguments{
\item{run}{= R6 Run object, one per simulation}

\item{lineage}{= Lineage object, to be removed from list of lineage pair choices}
}
\description{
This function removes lineage pairs in Run obj attr `choices`: list of lineage 
pair choices to coalesce.
}
