% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run.R
\name{Run}
\alias{Run}
\title{Run}
\description{
\code{Run} is an R6 class that is derived from \code{Model}.  It inherits
all simulation model features including CompartmentTypes, Compartments, 
and Lineages.  Unlike \code{Model}, a \code{Run} object is used to track 
an outcome of the simulation model and is therefore mutable where \code{Model}
is immutable.  For example, \code{Run} tracks which lineages are extant and 
the numbers of pairs of Lineages that can coalesce over time.  In addition, 
the number of unsampled compartments is stochastic outcome of the model that
is tracked at this level.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{extant.lineages}}{list of Lineage objects with sampling time t=0 (most recent)}

\item{\code{locations}}{a named list of Lineage objects keyed by Compartment}

\item{\code{counts}}{a data frame of population dynamics}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{extant.lineages}}{list of Lineage objects with sampling time t=0 (most recent)}

\item{\code{locations}}{a named list of Lineage objects keyed by Compartment}

\item{\code{counts}}{a data frame of population dynamics}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Run$new()}}
\item \href{#method-get.eventlog}{\code{Run$get.eventlog()}}
\item \href{#method-set.eventlog}{\code{Run$set.eventlog()}}
\item \href{#method-get.counts}{\code{Run$get.counts()}}
\item \href{#method-set.counts}{\code{Run$set.counts()}}
\item \href{#method-get.initial.conds}{\code{Run$get.initial.conds()}}
\item \href{#method-get.fixed.samplings}{\code{Run$get.fixed.samplings()}}
\item \href{#method-get.types}{\code{Run$get.types()}}
\item \href{#method-get.compartments}{\code{Run$get.compartments()}}
\item \href{#method-get.lineages}{\code{Run$get.lineages()}}
\item \href{#method-get.unsampled.hosts}{\code{Run$get.unsampled.hosts()}}
\item \href{#method-get.extant.lineages}{\code{Run$get.extant.lineages()}}
\item \href{#method-add.lineage}{\code{Run$add.lineage()}}
\item \href{#method-remove.lineage}{\code{Run$remove.lineage()}}
\item \href{#method-get.node.ident}{\code{Run$get.node.ident()}}
\item \href{#method-generate.unsampled}{\code{Run$generate.unsampled()}}
\item \href{#method-clear.unsampled}{\code{Run$clear.unsampled()}}
\item \href{#method-clone}{\code{Run$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$new(model)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.eventlog"></a>}}
\subsection{Method \code{get.eventlog()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$get.eventlog()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set.eventlog"></a>}}
\subsection{Method \code{set.eventlog()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$set.eventlog(e)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.counts"></a>}}
\subsection{Method \code{get.counts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$get.counts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set.counts"></a>}}
\subsection{Method \code{set.counts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$set.counts(df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.initial.conds"></a>}}
\subsection{Method \code{get.initial.conds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$get.initial.conds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.fixed.samplings"></a>}}
\subsection{Method \code{get.fixed.samplings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$get.fixed.samplings()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.types"></a>}}
\subsection{Method \code{get.types()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$get.types()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.compartments"></a>}}
\subsection{Method \code{get.compartments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$get.compartments()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.lineages"></a>}}
\subsection{Method \code{get.lineages()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$get.lineages()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.unsampled.hosts"></a>}}
\subsection{Method \code{get.unsampled.hosts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$get.unsampled.hosts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.extant.lineages"></a>}}
\subsection{Method \code{get.extant.lineages()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$get.extant.lineages(time, comp = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add.lineage"></a>}}
\subsection{Method \code{add.lineage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$add.lineage(lineage)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove.lineage"></a>}}
\subsection{Method \code{remove.lineage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$remove.lineage(lineage)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.node.ident"></a>}}
\subsection{Method \code{get.node.ident()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$get.node.ident(prefix = "Node")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate.unsampled"></a>}}
\subsection{Method \code{generate.unsampled()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$generate.unsampled(type.name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear.unsampled"></a>}}
\subsection{Method \code{clear.unsampled()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$clear.unsampled()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
