% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDynamics.R
\name{sim.dynamics}
\alias{sim.dynamics}
\title{sim.dynamics}
\usage{
sim.dynamics(mod, logfile = NULL, max.attempts = 3, chunk.size = 10000)
}
\arguments{
\item{mod:}{R6 object of class Model}

\item{logfile:}{str, path to write event log; if not specified, then 
this function will return a data.frame}

\item{max.attempts:}{int, number of attempts to simulate trajectories}

\item{chunk.size:}{int, number of rows to allocate to data.frame that 
stores events}
}
\value{
data.frame if user does not specify logfile
}
\description{
Simulate population trajectories for all compartments forward in time.
Results are recorded as a data frame or exported to a CSV file with the 
following information:
  time:  numeric, time of event in simulation time
  event:  str, type of event (birth, death, migration or transmission)
  from.comp:  str, Compartment originating the event
  to.comp:  str, for migration or transmission, the Compartment added to
  source:  str, for transmission only, compartment of the infecting host
}
\examples{
require(twt)
settings <- yaml.load_file("examples/SIRS_serial.yaml")
mod <- Model$new(settings)
sim.dynamics(mod, logfile="eventlog.csv")

}
