% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{sim.inner.tree}
\alias{sim.inner.tree}
\title{sim.inner.tree}
\usage{
sim.inner.tree(mod, e = NA)
}
\arguments{
\item{mod:}{R6 Model or Run object.  If user provides a Model object, then an 
EventLogger object, such as generated by eventlog.from.tree(), must also 
be provided.}

\item{e:}{(optional) R6 EventLogger object}
}
\value{
EventLogger object updated with inner tree events.
}
\description{
Simulate the coalescence of pathogen lineages within hosts (compartments)
AND resolve the migration events that may or may not involve sampled pathogen 
lineages within hosts.
}
\examples{
# load model
path <- system.file('extdata', 'SI.yaml', package='twt')

# load file and parse to construct MODEL object
settings <- yaml.load_file(path)
mod <- Model$new(settings)

# simulate outer tree - returns a Run object carrying EventLogger
run <- sim.outer.tree(mod)

# simulate inner tree
tree <- sim.inner.tree(run)
plot(tree)

}
