% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{sim.inner.tree}
\alias{sim.inner.tree}
\title{sim.inner.tree}
\usage{
sim.inner.tree(model, eventlog)
}
\arguments{
\item{model:}{R6 object from Model$new()}

\item{eventlog:}{R6 EventLogger object populated by sim.outer.tree() and sim.migrations()}
}
\value{
EventLogger object updated with inner tree events.
}
\description{
Simulate the coalescence of pathogen lineages within hosts (compartments)
AND resolve the migration events that may or may not involve sampled pathogen 
lineages within hosts.
}
\examples{
# load model
path <- system.file('extdata', 'SI.yaml', package='twt')

# load file and parse to construct MODEL object
settings <- yaml.load_file(path)
mod <- MODEL$new(settings)

# simulate outer tree
outer <- sim.outer.tree(mod)
tree <- sim.inner.tree(mod, outer)

}
