% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInnerTree.R
\name{sim.inner.tree}
\alias{sim.inner.tree}
\title{sim.inner.tree}
\usage{
sim.inner.tree(model, eventlog)
}
\arguments{
\item{model:}{R6 object from Model$new()}

\item{eventlog:}{R6 EventLogger object populated by sim.outer.tree() and sim.migrations()}
}
\description{
Simulate the coalescence of pathogen lineages within hosts (compartments)
AND resolve the migration events that may or may not involve sampled pathogen 
lineages within hosts.  EventLogger object updated with inner tree events.
}
\examples{


}
