% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{sim.outer.tree}
\alias{sim.outer.tree}
\title{sim.outer.tree}
\usage{
sim.outer.tree(model, eventlog)
}
\arguments{
\item{model:}{object of class 'MODEL'}

\item{eventlog:}{object of class 'EventLogger' to update}
}
\description{
After the objects are generated from user inputs ('loadInputs.R'), 
we need to initialize the list of fixed events.  We anticipate 
three use cases:
1. User provides a string or object representing the transmission tree
   that should be converted into an EventLogger object to use for
   inner tree simulation (`eventlog.from.tree`).
2. User manually inputs a host transmission tree into YAML format under 
   Compartments header (`init.branching.events`).
3. No host tree provided, transmission events need to be simulated under
   user-specified model (`sim.outer.tree`).
   
\code{sim.outer.tree} simulates transmission events and fixes them to the 
timeline of lineage sampled events.
}
\seealso{
init.branching.events, eventlog.from.tree
}
