% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{sim.outer.tree}
\alias{sim.outer.tree}
\title{sim.outer.tree}
\usage{
sim.outer.tree(mod, eventlog, chunk.size = 100)
}
\arguments{
\item{mod:}{R6 object of class Model}

\item{eventlog:}{character or data frame of event log from forward-time 
simulation of population trajectories}
}
\value{
data frame, containing subset of events that comprise
        the transmission tree
}
\description{
Given the events generated under the model by forward-time simulation in 
simulate.dynamics(), we determine the subset of events that are involved 
in the transmission tree relating the sampled hosts.  This will generally
be a much smaller subset of events.
}
\examples{
require(twt)
mod <- Model$new(yaml.load_file("examples/SIRS_serial.yaml"))
eventlog <- sim.dynamics(mod)
outer <- sim.outer.tree(mod, eventlog)
}
