% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOuterTree.R
\name{sim.outer.tree}
\alias{sim.outer.tree}
\title{sim.outer.tree}
\usage{
sim.outer.tree(model)
}
\arguments{
\item{model:}{object of class 'MODEL'}
}
\value{
object of class 'EventLogger'
}
\description{
After the objects are generated from user inputs ('loadInputs.R'), 
we need to initialize the list of fixed events.  We anticipate 
three use cases:
1. User provides a string or object representing the transmission tree
   that should be converted into an EventLogger object to use for
   inner tree simulation (`eventlog.from.tree`).
2. User manually inputs a host transmission tree into YAML format under 
   Compartments header (`init.branching.events`).
3. No host tree provided, transmission events need to be simulated under
   user-specified model (`sim.outer.tree`).
   
\code{sim.outer.tree} simulates transmission, migration and transition events 
and fixes them to the timeline of lineage sampled events.
}
\details{
A *transmission* is the passage of one or more Lineages from a source 
Compartment to an uninfected recipient Compartment that did not carry any 
Lineages.  

A *migration* is the passage of one or more Lineages from a source Compartment
to a recipient Compartment that has already been infected (it already carries
one or more Lineages).

A *transition* occurs when a Compartment switches to another Type.  This 
is useful for handling host death events, for example.
}
\examples{

#' # load susceptible-infected (SI) compartmental model
path <- system.file('extdata', 'structSI.yaml', package='twt')
settings <- yaml.load_file(path)
model <- Model$new(settings)

run <- sim.outer.tree(model)
run$get.eventlog()

}
\seealso{
init.branching.events, eventlog.from.tree
}
